requirements.txt
s3fs
python-dotenv
chromadb
streamlit
langchain-core
langchain-openai
langchain
langchain-community
